requirements.txt
streamlit
google-generativeai
python-dotenv
tavily-python
pillow
pyttsx3